(PC-26154)[API] refactor: move Address & OffererAddrem from geography module to offerer module
