+++ api/src/pcapi/core/offers/models.py
@@ -44,8 +44,8 @@ if typing.TYPE_CHECKING:
     from pcapi.core.criteria.models import Criterion
     from pcapi.core.educational.models import CollectiveOffer
     from pcapi.core.educational.models import CollectiveOfferTemplate
-    from pcapi.core.geography.models import OffererAddress
     from pcapi.core.offerers.models import Offerer
+    from pcapi.core.offerers.models import OffererAddress
     from pcapi.core.offerers.models import Venue
     from pcapi.core.providers.models import Provider
     from pcapi.core.users.models import User
